refactor(GtfsFieldSelector): Extract GtfsFieldSelector component
