+++ lib/manager/components/transform/NormalizeField.js
@@ -13,12 +13,11 @@ import {
 } from 'react-bootstrap'
 import Select from 'react-select'
 
-import { getGtfsSpec } from '../../../common/util/config'
+import GtfsFieldSelector from '../GtfsFieldSelector'
 import SubstitutionRow from './SubstitutionRow'
 import Transform from './Transform'
 
 import type {
-  GtfsSpecTable,
   NormalizeFieldFields,
   ReactSelectOption,
   Substitution
@@ -124,8 +123,8 @@ export default class NormalizeField extends Transform<NormalizeFieldFields> {
   render () {
     const { transformation } = this.props
     const {
-      capitalize,
       capitalizationExceptions,
+      capitalize,
       fieldName,
       performSubstitutions,
       substitutions
@@ -268,49 +267,4 @@ export default class NormalizeField extends Transform<NormalizeFieldFields> {
       </div>
     )
   }
-}
-
-type GtfsFieldSelectorProps = {
-  onChange: (ReactSelectOption => void),
-  selectedField: ?string,
-  tableName: ?string
-}
-
-/**
- * Obtains the desired GTFS spec table.
- */
-function getGtfsTableSpec (tableName: string): ?GtfsSpecTable {
-  const spec = getGtfsSpec()
-  if (!spec) {
-    throw new Error('GTFS spec could not be found!')
-  }
-  const tableFile = `${tableName}.txt`
-  return spec.find(t => t.name === tableFile)
-}
-
-/**
- * Dropdown selector to select a field from a given GTFS table.
- */
-const GtfsFieldSelector = (props: GtfsFieldSelectorProps) => {
-  const { onChange, selectedField, tableName } = props
-  let options = []
-  let placeholder = 'First, select a table above'
-  if (tableName) {
-    const table = getGtfsTableSpec(tableName)
-    options = table && table.fields.map(field => ({ label: field.name, value: field.name }))
-    placeholder = 'Choose the field to normalize'
-  }
-
-  return (
-    <Select
-      clearable={false}
-      disabled={!tableName}
-      onChange={onChange}
-      options={options}
-      placeholder={placeholder}
-      searchable
-      style={{margin: '10px 0px', width: '230px'}}
-      value={selectedField}
-    />
-  )
 }
